refactor: feature/auth 코드 수정 및 적용
