+++ frontend/src/features/auth/api/auth.api.ts
@@ -1,5 +1,4 @@
-import { IUserLogin, IUserRegistration } from '@/models/user.model';
-import { httpClient } from './http';
+import { httpClient } from '@/shared/utils/http';
 
 interface IAuthProps {
   email?: string;
@@ -15,16 +14,19 @@ export const checkedNickname = async (nickname: IAuthProps) => {
   return await httpClient.post('/users/checkedNickname', nickname);
 };
 
-export const join = async (
-  userData: Omit<IUserRegistration, 'password_checked'>,
-) => {
+export const join = async (userData: Omit<IAuthProps, 'password_checked'>) => {
   return await httpClient.post('/users/join', userData);
 };
 
-export const login = async (userData: IUserLogin) => {
+export const login = async (userData: IAuthProps) => {
   return await httpClient.post('/users/login', userData);
 };
 
 export const logout = async () => {
   return await httpClient.post('/users/logout');
+};
+
+export const getUserInfo = async () => {
+  const { data } = await httpClient.get('/users/me');
+  return data;
 };
