--- frontend/src/features/auth/api/user.api.ts
@@ -1,6 +1,0 @@
-import { httpClient } from '../../../shared/utils/http';
-
-export const getUserInfo = async () => {
-  const { data } = await httpClient.get('/users/me');
-  return data;
-};
