--- frontend/src/models/user.model.ts
@@ -1,14 +1,0 @@
-export interface IUser {
-  email: string;
-  nickname: string;
-}
-
-export interface IUserLogin {
-  email: string;
-  password: string;
-}
-
-export interface IUserRegistration extends IUser {
-  password: string;
-  password_checked: string;
-}
